+++ packages/frontend/src/hooks/cartesianChartConfig/useCartesianChartConfig.mock.ts
@@ -1,5 +1,6 @@
 import {
     CartesianSeriesType,
+    ChartType,
     CompiledDimension,
     CompiledTable,
     DimensionType,
@@ -98,7 +99,7 @@ export const expectedSimpleSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -114,7 +115,7 @@ export const expectedSimpleSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -138,7 +139,7 @@ export const expectedPivotedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -155,7 +156,7 @@ export const expectedPivotedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -172,7 +173,7 @@ export const expectedPivotedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -189,7 +190,7 @@ export const expectedPivotedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -251,6 +252,7 @@ export const expectedMultiPivotedSeriesMap: Record<string, Series> = {
         smooth: undefined,
         areaStyle: undefined,
         stack: undefined,
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -270,7 +272,7 @@ export const expectedMultiPivotedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -289,8 +291,8 @@ export const expectedMultiPivotedSeriesMap: Record<string, Series> = {
         areaStyle: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
         stack: undefined,
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -310,7 +312,7 @@ export const expectedMultiPivotedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -333,7 +335,7 @@ export const existingMixedSeries: Series[] = [
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -349,7 +351,7 @@ export const existingMixedSeries: Series[] = [
         areaStyle: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         stack: undefined,
         encode: {
             xRef: {
@@ -371,7 +373,7 @@ export const expectedMixedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -387,7 +389,7 @@ export const expectedMixedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -403,7 +405,7 @@ export const expectedMixedSeriesMap: Record<string, Series> = {
         areaStyle: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         stack: undefined,
         encode: {
             xRef: {
@@ -421,7 +423,7 @@ export const expectedMixedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -438,7 +440,7 @@ export const expectedMixedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -455,7 +457,7 @@ export const expectedMixedSeriesMap: Record<string, Series> = {
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -474,7 +476,7 @@ export const mergedMixedSeries: Series[] = [
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -491,7 +493,7 @@ export const mergedMixedSeries: Series[] = [
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -509,7 +511,7 @@ export const mergedMixedSeries: Series[] = [
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -526,7 +528,7 @@ export const mergedMixedSeries: Series[] = [
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -543,7 +545,7 @@ export const mergedMixedSeries: Series[] = [
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -560,7 +562,7 @@ export const mergedMixedSeries: Series[] = [
         stack: undefined,
         showSymbol: undefined,
         smooth: undefined,
-
+        yAxisIndex: 0,
         encode: {
             xRef: {
                 field: 'my_dimension',
@@ -590,3 +592,134 @@ export const groupedMixedSeries: Array<{ index: number; value: Series[] }> = [
         value: [mergedMixedSeries[5]],
     },
 ];
+
+export const useCartesianChartConfigParamsMock = {
+    pivotKeys: undefined,
+    setPivotDimensions: () => {},
+    chartType: ChartType.CARTESIAN,
+    initialChartConfig: {
+        layout: {
+            xField: 'orders_customer_id',
+            yField: ['orders_total_order_amount', 'orders_fulfillment_rate'],
+        },
+        eChartsConfig: {
+            series: [],
+        },
+    },
+    resultsData: {
+        rows: [],
+        metricQuery: {
+            dimensions: ['orders_customer_id'],
+            metrics: ['orders_total_order_amount', 'orders_fulfillment_rate'],
+            filters: {},
+            sorts: [
+                {
+                    fieldId: 'orders_total_order_amount',
+                    descending: true,
+                },
+            ],
+            limit: 500,
+            tableCalculations: [],
+            additionalMetrics: [],
+        },
+    },
+    columnOrder: [
+        'orders_customer_id',
+        'orders_total_order_amount',
+        'orders_fulfillment_rate',
+    ],
+    explore: {
+        name: 'orders',
+        tags: [],
+        label: 'Orders',
+        tables: {
+            orders: {
+                name: 'orders',
+                label: 'Orders',
+                schema: 'jaffle',
+                metrics: {
+                    fulfillment_rate: {
+                        sql: 'CASE WHEN ${is_completed} THEN 1 ELSE 0 END',
+                        name: 'fulfillment_rate',
+                        type: 'average',
+                        index: 2,
+                        label: 'Fulfillment rate',
+                        round: 1,
+                        table: 'orders',
+                        format: 'percent',
+                        hidden: false,
+                        filters: [],
+                        fieldType: 'metric',
+                        tableLabel: 'Orders',
+                        compiledSql:
+                            'AVG(CASE WHEN ("orders".is_completed) THEN 1 ELSE 0 END)',
+                        description: 'Average of Is completed',
+                        isAutoGenerated: false,
+                        tablesReferences: ['orders'],
+                        showUnderlyingValues: [
+                            'amount',
+                            'customers.first_name',
+                        ],
+                    },
+
+                    total_order_amount: {
+                        sql: '${TABLE}.amount',
+                        name: 'total_order_amount',
+                        type: 'sum',
+                        index: 6,
+                        label: 'Total order amount',
+                        round: 2,
+                        table: 'orders',
+                        format: 'usd',
+                        hidden: false,
+                        filters: [],
+                        fieldType: 'metric',
+                        tableLabel: 'Orders',
+                        compiledSql: 'SUM("orders".amount)',
+                        description: 'Sum of Amount',
+                        isAutoGenerated: false,
+                        tablesReferences: ['orders'],
+                    },
+                },
+                database: 'postgres',
+                sqlTable: '"postgres"."jaffle"."orders"',
+                dimensions: {
+                    customer_id: {
+                        sql: '${TABLE}.customer_id',
+                        name: 'customer_id',
+                        type: 'number',
+                        index: 2,
+                        label: 'Customer id',
+                        table: 'orders',
+                        hidden: false,
+                        fieldType: 'dimension',
+                        tableLabel: 'Orders',
+                        compiledSql: '"orders".customer_id',
+                        description: 'Foreign key to the customers table',
+                        tablesReferences: ['orders'],
+                    },
+                },
+                description:
+                    'This table has basic information about orders, as well as some derived facts based on payments',
+                lineageGraph: {
+                    orders: [
+                        {
+                            name: 'stg_orders',
+                            type: 'model',
+                        },
+                        {
+                            name: 'stg_payments',
+                            type: 'model',
+                        },
+                    ],
+                    stg_orders: [],
+                    stg_payments: [],
+                },
+                orderFieldsBy: 'LABEL',
+            },
+        },
+        baseTable: 'orders',
+        joinedTables: [],
+        targetDatabase: 'postgres',
+    },
+};