+++ packages/frontend/src/hooks/cartesianChartConfig/useCartesianChartConfig.test.ts
@@ -1,3 +1,6 @@
+import { CartesianSeriesType } from '@lightdash/common';
+import { renderHook } from '@testing-library/react-hooks';
+import useCartesianChartConfig from './useCartesianChartConfig';
 import {
     existingMixedSeries,
     expectedMixedSeriesMap,
@@ -10,6 +13,7 @@ import {
     multiPivotSeriesMapArgs,
     pivotSeriesMapArgs,
     simpleSeriesMapArgs,
+    useCartesianChartConfigParamsMock,
 } from './useCartesianChartConfig.mock';
 import {
     getExpectedSeriesMap,
@@ -174,5 +178,66 @@ describe('getSeriesGroupedByField', () => {
         expect(
             getSeriesGroupedByField(Object.values(mergedMixedSeries)),
         ).toStrictEqual(groupedMixedSeries);
+    });
+});
+
+describe('useCartesianChartConfig', () => {
+    test('should default series yAxisIndex to 0', () => {
+        const { result } = renderHook(
+            // @ts-expect-error partially mock params for hook
+            () => useCartesianChartConfig(useCartesianChartConfigParamsMock),
+        );
+
+        const series =
+            result.current.validCartesianConfig!.eChartsConfig.series!;
+
+        series.forEach((serie) => expect(serie.yAxisIndex).toBe(0));
+    });
+    test('should set undefined yAxisIndex to 0', () => {
+        const seriesFromOldChart = [
+            {
+                type: CartesianSeriesType.BAR,
+                encode: {
+                    xRef: {
+                        field: 'orders_customer_id',
+                    },
+                    yRef: {
+                        field: 'orders_total_order_amount',
+                    },
+                },
+                yAxisIndex: 1,
+            },
+            {
+                type: CartesianSeriesType.BAR,
+                encode: {
+                    xRef: {
+                        field: 'orders_customer_id',
+                    },
+                    yRef: {
+                        field: 'orders_fulfillment_rate',
+                    },
+                },
+            },
+        ];
+
+        const { result } = renderHook(() =>
+            // @ts-expect-error partially mock params for hook
+            useCartesianChartConfig({
+                ...useCartesianChartConfigParamsMock,
+                initialChartConfig: {
+                    ...useCartesianChartConfigParamsMock.initialChartConfig,
+
+                    eChartsConfig: {
+                        series: seriesFromOldChart,
+                    },
+                },
+            }),
+        );
+
+        const series =
+            result.current.validCartesianConfig!.eChartsConfig.series!;
+
+        expect(series[0].yAxisIndex).toBe(1);
+        expect(series[1].yAxisIndex).toBe(0);
     });
 });